+++ src/base/widget_consts.h
@@ -1068,6 +1068,12 @@ BEGIN_C_DECLS
 #define WIDGET_PROP_ANIMATE_ANIMATING_TIME "animate:animating_time"
 
 /**
+ * @const WIDGET_PROP_DIRTY_RECT
+ * 控件脏矩形区域。
+ */
+#define WIDGET_PROP_DIRTY_RECT  "dirty_rect"
+
+/**
  * @enum widget_type_t
  * @annotation ["scriptable", "string"]
  * @prefix WIDGET_TYPE_